# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.CloudRunServiceDetails:
    properties:
      created_time:
        type: string
      image:
        type: string
      location:
        type: string
      name:
        type: string
      scaling:
        $ref: '#/definitions/api.ServiceScaling'
      status:
        type: string
      updated_time:
        type: string
      url:
        type: string
    type: object
  api.PaginatedDeploymentsResponse:
    properties:
      count:
        type: integer
      deployments:
        items:
          $ref: '#/definitions/models.Deployment'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
    type: object
  api.RequestBody:
    properties:
      container_image:
        type: string
      max_instances:
        type: integer
      min_instances:
        type: integer
      name:
        type: string
      port:
        type: integer
    type: object
  api.ServiceScaling:
    properties:
      max_instances:
        type: integer
      min_instances:
        type: integer
    type: object
  models.Deployment:
    properties:
      container_image:
        type: string
      created_at:
        type: string
      id:
        type: string
      max_instances:
        type: integer
      min_instances:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_email:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A REST API for managing Cloud Run deployments and container images
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 0p5dev Controller API
  version: "1.0"
paths:
...
  /deployments:
    get:
      description: Get a paginated list of deployments for the authenticated user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Search in name, url, and container_image
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of deployments
          schema:
            $ref: '#/definitions/api.PaginatedDeploymentsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve deployments
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List deployments
      tags:
      - deployments
    put:
      consumes:
      - application/json
      description: Deploy a container image to Cloud Run
      parameters:
      - description: Deployment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Deployment successful with service URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "408":
          description: Deployment cancelled by client
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Deployment failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new deployment
      tags:
      - deployments
  /deployments/{name}:
    delete:
      description: Delete a Cloud Run deployment and remove it from the database
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Deployment name is required
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete deployment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a deployment
      tags:
      - deployments
    get:
      description: Retrieve detailed information about a specific deployment
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment details
          schema:
            $ref: '#/definitions/api.CloudRunServiceDetails'
        "400":
          description: Deployment name is required
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve deployment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get deployment by name
      tags:
      - deployments
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"